+++ src/com/main/FastEnemy.java
@@ -16,11 +16,11 @@ public class FastEnemy extends GameObject {
 	}
 
 	/*
-	 * Setting the boundaries of the Basic enemy
+	 * Setting the boundaries of the Smart enemy
 	 */
 
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int)x, (int)y, 16, 16);
 	}
 
 	@Override
@@ -40,7 +40,7 @@ public class FastEnemy extends GameObject {
 	@Override
 	public void render(Graphics g) {
 		g.setColor(Color.PINK);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 
 	}
