+++ src/com/main/Trail.java
@@ -16,7 +16,7 @@ public class Trail extends GameObject{
 	
 	private int width, height;
 	
-	public Trail(int x, int y, ID id, Color color, int width, int height, float life, Handler handler) {
+	public Trail(float x, float y, ID id, Color color, int width, int height, float life, Handler handler) {
 		super(x, y, id);
 		this.handler = handler;
 		this.color=color;
@@ -39,7 +39,7 @@ public class Trail extends GameObject{
 		g2d.setComposite(makeTransparent(alpha));
 		
 		g.setColor(color);
-		g.fillRect(x,y,width, height);
+		g.fillRect((int)x,(int)y,width, height);
 		
 	}
